requirements.txt
PyPDF2
google-generativeai
python-dotenv
fastapi
uvicorn